setup: Fix dependency handling
In the present configuration a downstream user of this library is
left with the responsibility of shipping and installing the
``pbr`` dependency.

Since shellvars is the one requiring it I think it would be
pertinent of it to actually ask for it's installation.

Both ``setup_requires`` and ``install_requires`` has to be set for
this to happen [0].

0: https://setuptools.readthedocs.io/en/latest/setuptools.html#new-and-changed-setup-keywords